+++ src/components/DecoratedText/DecoratedText.test.tsx
@@ -1,8 +1,6 @@
 import React from "react";
 import { create } from "react-test-renderer";
 import { DecoratedText } from "./DecoratedText";
-import { HashMap } from "@rimbu/hashed";
-import { Char, CharMetadata } from "./Metadata";
 import { Pinyin } from "@/glyphs";
 
 describe("<DecoratedText>", () => {
@@ -11,10 +9,9 @@ describe("<DecoratedText>", () => {
       const renderer = create(
         <DecoratedText
           text={"我看书"}
-          metadataByChar={HashMap.of<Char, CharMetadata>([
-            "书",
-            { annotation: "shū", className: Pinyin.flat }
-          ])}
+          metadataByChar={{
+            书: { annotation: "shū", className: Pinyin.flat }
+          }}
         />
       );
 
@@ -25,10 +22,7 @@ describe("<DecoratedText>", () => {
   describe("when receiving text with leading and trailing space and no metadata", () => {
     it("should trim the text and render just plain characters in a single paragraph", () => {
       const renderer = create(
-        <DecoratedText
-          text={"  \t \nDodo  \r\n  \t"}
-          metadataByChar={HashMap.empty()}
-        />
+        <DecoratedText text={"  \t \nDodo  \r\n  \t"} metadataByChar={{}} />
       );
 
       expect(renderer.toJSON()).toMatchSnapshot();
@@ -38,10 +32,7 @@ describe("<DecoratedText>", () => {
   describe("when receiving text with multiple intermediate carriageReturn + lineFeed", () => {
     it("should ignore carriage returns", () => {
       const renderer = create(
-        <DecoratedText
-          text={"Alpha\r\nBeta\r\nGamma"}
-          metadataByChar={HashMap.empty()}
-        />
+        <DecoratedText text={"Alpha\r\nBeta\r\nGamma"} metadataByChar={{}} />
       );
 
       expect(renderer.toJSON()).toMatchSnapshot();
@@ -53,7 +44,7 @@ describe("<DecoratedText>", () => {
       const renderer = create(
         <DecoratedText
           text={"Alpha\n\n\n\nBeta\n\n\n\n\n\n\nGamma"}
-          metadataByChar={HashMap.empty()}
+          metadataByChar={{}}
         />
       );
 
@@ -66,27 +57,20 @@ describe("<DecoratedText>", () => {
       const renderer = create(
         <DecoratedText
           text={"Alpha1\nAlpha2\n\nBeta1\n我看书\nBeta3\nBeta4\n\n\n\nGamma"}
-          metadataByChar={HashMap.of<Char, CharMetadata>(
-            [
-              "我",
-              {
-                annotation: "wǒ",
-                className: Pinyin.fallingRising
-              }
-            ],
-            [
-              "看",
-              {
-                annotation: "kàn"
-              }
-            ],
-            [
-              "书",
-              {
-                className: Pinyin.flat
-              }
-            ]
-          )}
+          metadataByChar={{
+            我: {
+              annotation: "wǒ",
+              className: Pinyin.fallingRising
+            },
+
+            看: {
+              annotation: "kàn"
+            },
+
+            书: {
+              className: Pinyin.flat
+            }
+          }}
         />
       );