+++ src/components/DecoratedText/index.ts
@@ -1,3 +1,1 @@
 export * from "./DecoratedText";
-
-export * from "./Metadata";